refactor: 💡 Move env variable test fixture to global setup file
✅ Closes: https://hashicorp.atlassian.net/browse/ICU-14777
