+++ e2e-tests/admin/playwright.config.js
@@ -3,11 +3,11 @@
  * SPDX-License-Identifier: BUSL-1.1
  */
 
-import { defineConfig, devices, test as baseTest } from '@playwright/test';
+import { defineConfig, devices } from '@playwright/test';
 
 /** @type {import('@playwright/test').PlaywrightTestConfig} */
 export default defineConfig({
-  globalSetup: './global-setup',
+  globalSetup: '../global-setup',
   outputDir: './artifacts/test-failures',
   timeout: 90000, // Each test is given 90s to complete
   workers: 1, // Tests need to be run in serial, otherwise there may be conflicts when using the CLI
@@ -34,36 +34,4 @@ export default defineConfig({
       use: { ...devices['Desktop Safari'] },
     },
   ],
-});
-
-export const test = baseTest.extend({
-  adminAuthMethodId: process.env.E2E_PASSWORD_AUTH_METHOD_ID,
-  adminLoginName: process.env.E2E_PASSWORD_ADMIN_LOGIN_NAME,
-  adminPassword: process.env.E2E_PASSWORD_ADMIN_PASSWORD,
-  awsAccessKeyId: process.env.E2E_AWS_ACCESS_KEY_ID,
-  awsBucketName: process.env.E2E_AWS_BUCKET_NAME,
-  awsHostSetFilter: process.env.E2E_AWS_HOST_SET_FILTER,
-  awsHostSetIps: process.env.E2E_AWS_HOST_SET_IPS,
-  awsRegion: process.env.E2E_AWS_REGION,
-  awsSecretAccessKey: process.env.E2E_AWS_SECRET_ACCESS_KEY,
-  bucketAccessKeyId: process.env.E2E_BUCKET_ACCESS_KEY_ID,
-  bucketEndpointUrl: process.env.E2E_BUCKET_ENDPOINT_URL,
-  bucketName: process.env.E2E_BUCKET_NAME,
-  bucketSecretAccessKey: process.env.E2E_BUCKET_SECRET_ACCESS_KEY,
-  ldapAddr: process.env.E2E_LDAP_ADDR,
-  ldapAdminDn: process.env.E2E_LDAP_ADMIN_DN,
-  ldapAdminPassword: process.env.E2E_LDAP_ADMIN_PASSWORD,
-  ldapDomainDn: process.env.E2E_LDAP_DOMAIN_DN,
-  ldapGroupName: process.env.E2E_LDAP_GROUP_NAME,
-  ldapUserName: process.env.E2E_LDAP_USER_NAME,
-  ldapUserPassword: process.env.E2E_LDAP_USER_PASSWORD,
-  region: process.env.E2E_REGION,
-  sshCaKey: process.env.E2E_SSH_CA_KEY,
-  sshCaKeyPublic: process.env.E2E_SSH_CA_KEY_PUBLIC,
-  sshKeyPath: process.env.E2E_SSH_KEY_PATH,
-  sshUser: process.env.E2E_SSH_USER,
-  targetAddress: process.env.E2E_TARGET_ADDRESS,
-  targetPort: process.env.E2E_TARGET_PORT,
-  vaultAddr: process.env.E2E_VAULT_ADDR,
-  workerTagEgress: process.env.E2E_WORKER_TAG_EGRESS,
 });
