+++ e2e-tests/admin/tests/target-ent.spec.js
@@ -3,10 +3,9 @@
  * SPDX-License-Identifier: BUSL-1.1
  */
 
-import { test } from '../playwright.config.js';
+import { test, authenticatedState } from '../../global-setup.js';
 import { expect } from '@playwright/test';
 
-import { authenticatedState } from '../../global-setup.js';
 import {
   authenticateBoundaryCli,
   checkBoundaryCli,
