+++ e2e-tests/admin/tests/worker-tags.spec.js
@@ -3,11 +3,10 @@
  * SPDX-License-Identifier: BUSL-1.1
  */
 
-import { test } from '../playwright.config.js';
+import { test, authenticatedState } from '../../global-setup.js';
 import { expect } from '@playwright/test';
 import { customAlphabet } from 'nanoid';
 
-import { authenticatedState } from '../../global-setup.js';
 import { WorkersPage } from '../pages/workers.js';
 
 test.use({ storageState: authenticatedState });
