+++ e2e-tests/global-setup.js
@@ -7,6 +7,7 @@ import { chromium } from '@playwright/test';
 import { checkEnv } from './helpers/general.js';
 
 import { LoginPage } from './admin/pages/login.js';
+import { test as baseTest } from 'playwright/types/test';
 
 async function globalSetup() {
   await checkEnv([
@@ -41,3 +42,35 @@ async function globalSetup() {
 export default globalSetup;
 
 export const authenticatedState = './admin/artifacts/authenticated-state.json';
+
+export const test = baseTest.extend({
+  adminAuthMethodId: process.env.E2E_PASSWORD_AUTH_METHOD_ID,
+  adminLoginName: process.env.E2E_PASSWORD_ADMIN_LOGIN_NAME,
+  adminPassword: process.env.E2E_PASSWORD_ADMIN_PASSWORD,
+  awsAccessKeyId: process.env.E2E_AWS_ACCESS_KEY_ID,
+  awsBucketName: process.env.E2E_AWS_BUCKET_NAME,
+  awsHostSetFilter: process.env.E2E_AWS_HOST_SET_FILTER,
+  awsHostSetIps: process.env.E2E_AWS_HOST_SET_IPS,
+  awsRegion: process.env.E2E_AWS_REGION,
+  awsSecretAccessKey: process.env.E2E_AWS_SECRET_ACCESS_KEY,
+  bucketAccessKeyId: process.env.E2E_BUCKET_ACCESS_KEY_ID,
+  bucketEndpointUrl: process.env.E2E_BUCKET_ENDPOINT_URL,
+  bucketName: process.env.E2E_BUCKET_NAME,
+  bucketSecretAccessKey: process.env.E2E_BUCKET_SECRET_ACCESS_KEY,
+  ldapAddr: process.env.E2E_LDAP_ADDR,
+  ldapAdminDn: process.env.E2E_LDAP_ADMIN_DN,
+  ldapAdminPassword: process.env.E2E_LDAP_ADMIN_PASSWORD,
+  ldapDomainDn: process.env.E2E_LDAP_DOMAIN_DN,
+  ldapGroupName: process.env.E2E_LDAP_GROUP_NAME,
+  ldapUserName: process.env.E2E_LDAP_USER_NAME,
+  ldapUserPassword: process.env.E2E_LDAP_USER_PASSWORD,
+  region: process.env.E2E_REGION,
+  sshCaKey: process.env.E2E_SSH_CA_KEY,
+  sshCaKeyPublic: process.env.E2E_SSH_CA_KEY_PUBLIC,
+  sshKeyPath: process.env.E2E_SSH_KEY_PATH,
+  sshUser: process.env.E2E_SSH_USER,
+  targetAddress: process.env.E2E_TARGET_ADDRESS,
+  targetPort: process.env.E2E_TARGET_PORT,
+  vaultAddr: process.env.E2E_VAULT_ADDR,
+  workerTagEgress: process.env.E2E_WORKER_TAG_EGRESS,
+});
